Fixed exists() and starting to clean architecture

+++ TP03/copy.c
@@ -66,7 +66,7 @@ int copy(const char* from, const char* to) {
         } while (readNb > 0);
     }
     if (readNb < 0) {
-        hdlReadErr(from, 0, 0, NULL);
+        hdlReadErr(from, 0, 0, (int)NULL);
         return hdlCopyErr(from, to, 1, 0, fd_from, fd_to);
     }
     if (close(fd_to) < 0) {
